pyproject.toml
[project]
name = "aps_digitaltwin"
description = "APS Digital Twin package"
version = "0.1.0"
authors = [
    { name = "Richard Somers", email = "rsomers@sheffield.ac.uk" }
]
dependencies = [
    "matplotlib~=3.5.1",
    "numpy~=1.22.4",
    "pandas~=1.3.5",
    "pygad~=2.18.1",
    "pyswarms~=1.3.0",
    "python-dotenv~=1.0.0",
]

[build-system]
requires = [
    "setuptools>=64",
    "setuptools_scm[toml]>=6.2",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["scripts"]
